New classes: Character and BattleTheme, and adding some descriptions to then
